refact: UserServlet
같은 경로 내의 처리 일관화를 위해 하나의 서블릿으로 변경 후 if 문을 통해 경로별 분기처리
